requirements.txt
fastavro
csv
json
pyorc
pandas
yfinance
os
bs4
kafka-python
requests
